Realtime Source Code
This is a CIN.  Start the review at the CINRun Main Function definition.  This is the entry point.  There is a lot of code here but it basically reads a formula in as a string.  Then it parses it out and fills in the formula variables (from test results, constants used in test, and so on).  Does the math following the order of operations, compares the results to the limits and reports them back to the test.